==> 5_VRD_save_Horizontal_and_Vertical_angle_of_Exit_and_Entry_points.ipynb <==
# Save the vertical and horizontal angle of exit and entry point with respect to the nest

#### The data is saved in recording floder in bee_id floder

import os
import pandas as pd
import numpy as np
import math

# matplot
import matplotlib.pyplot as plt

from matplotlib.projections import PolarAxes
from mpl_toolkits.axisartist.floating_axes import GridHelperCurveLinear, FloatingSubplot
import mpl_toolkits.axisartist.grid_finder as gf

# Config
import configparser

bee_id = '2R'
recording = '20210414_163606'

common_path = 'path'
recording_condition = '{0}_learn'.format(recording)
recording_path = os.path.join(common_path,bee_id,recording_condition)


recording_path

# get the clutter data
clutter_arena_data = pd.read_csv('{}/clutter_arena_data.csv'.format(common_path))
clutter_arena_data

# nest location
nest_loc = list([clutter_data.nest_loc_x[0],clutter_data.nest_loc_y[0]])
othrer_point = list([clutter_data.other_pt_x[0],clutter_data.other_pt_y[0]])
print(nest_loc)
print(othrer_point)

#### Getting the exit points

# getting the exit point data
exit_mid_mean = pd.read_csv('{0}/exitmid_mean_df_{1}.csv'.format(recording_path,recording_condition),
                       header=[0]) # in_mid_out direction data
exit_mid_mean.head()

#### Getting the entry points

# getting the entry point data
entry_mid_mean = pd.read_csv('{0}/entrymid_mean_df_{1}.csv'.format(recording_path,recording_condition),
                       header=[0]) # out_mid_in direction data
entry_mid_mean.head()

# Getting actual distance, horizontal angle, vertical angle of exit and entry points


def get_dist_angle_HorzVert (mid_mean_df, nest_loc,othrer_point):
    
    ''' The function calculates the dsitance in 3D, horizontal angle and vertical angle'''
    
    Point_actDist = list() # exit point actual distance
    Point_horzAng_deg = list() # exit point horizontal angle
    Point_vertAng_deg = list() #exit point vertical angle
    x_coord = list(mid_mean_df.x_mean)
    y_coord = list(mid_mean_df.y_mean)
    z_coord = list(mid_mean_df.z_mean)

    for ii in range(len(x_coord)):
        dist = np.sqrt((nest_loc[0]-x_coord[ii])**2 + 
                            (nest_loc[1]-y_coord[ii])**2 + 
                            (z_coord[ii]**2))
        # Horizontal angle

    horz_angle_rad = (np.arctan2((x_coord[ii]-nest_loc[0]),(y_coord[ii]-nest_loc[1])) - 
                                np.arctan2((othrer_point[0]-nest_loc[0]),(othrer_point[1]-nest_loc[1])))
        horz_angle_deg = np.rad2deg(horz_angle_rad % (2 * np.pi))
# --------------------------------------alternative ----------------------------------------------------------

        # vertical angle
        vert_angle_deg = np.degrees(np.arcsin(z_coord[ii]/dist))

        Point_actDist.append(dist)
        Point_horzAng_deg.append(horz_angle_deg)
        Point_vertAng_deg.append(vert_angle_deg)
        
    return (Point_actDist,Point_horzAng_deg,Point_vertAng_deg)

exitPoint_actDist,exitPoint_horzAng,exitPoint_vertAng = get_dist_angle_HorzVert(exit_mid_mean,nest_loc,othrer_point)

print(exitPoint_actDist[0:5])
print(exitPoint_horzAng[0:5])
print(exitPoint_vertAng[0:5])

entryPoint_actDist, entryPoint_horzAng, entryPoint_vertAng = get_dist_angle_HorzVert(entry_mid_mean,nest_loc,othrer_point)

print(entryPoint_actDist[0:5])
print(entryPoint_horzAng[0:5])
print(entryPoint_vertAng[0:5])

#### Ploting the histogram 

# ploting hozrizontal angle out of 360

plt.hist(exitPoint_horzAng, bins=90)

# fig1, ax = plt.subplots(figsize=(10,6))
# ax.hist(exitPoint_horzAng, bins=90)
# ax.set_xlabel('Horizontal angle')
# ax.set_ylabel('Exit points')
# ax.set_title('Horizontal angle of exit points')
# plt.xlim([0,360])

# plot entry horizontal angle 

plt.hist(entryPoint_horzAng)

# fig2, ax = plt.subplots(figsize=(10,6))
# ax.hist(entryPoint_horzAng, bins=90)

# ax.set_xlabel('Horizontal angle')
# ax.set_ylabel('Entry points')
# ax.set_title('Horizontal angle of entry points')
# plt.xlim([0,360])


plt.hist(exitPoint_vertAng)

# plot entry point vertical angles

plt.hist(entryPoint_vertAng)

# fig4, ax = plt.subplots(figsize=(10,6))
# ax.hist(entryPoint_vertAng, bins=90)
# ax.set_xlabel('Vertical angle')
# ax.set_ylabel('Entry points')
# ax.set_title('Vertical angle entry point')
# plt.xlim([0,90])


clutterDsitAng_data = pd.read_csv('{}/clutterDsitAng_data.csv'.format(common_path))
clutterDsitAng_data.head()

clutter_horzAng = list(clutterDsitAng_data.horz_angle)
clutter_vertAng = list(clutterDsitAng_data.vert_angle)

### vert angle vs horz angle

# plotting vertical angle against horizontal angle
fig5, ax = plt.subplots(figsize=(11,8))
ax.scatter(exitPoint_horzAng,exitPoint_vertAng,label = 'Exit')
ax.scatter(entryPoint_horzAng,entryPoint_vertAng, label = 'Entry')
ax.set_xlabel('Horizontal angle')
ax.set_ylabel('Vertical angle')
ax.set_title('Vertical Vs Horizontal angle_exit/entry points_{0}_{1}'.format(bee_id,recording_condition))
# plot the line at cylinder angle 
ax.scatter(clutter_horzAng,clutter_vertAng,color = 'green')
ax.legend()

# both have same length, in case not
exit_annot_list = list(range(len(exitPoint_horzAng)))
entry_annot_list = list(range(len(entryPoint_horzAng)))

# lableling
label = 'corner angle'
ax.text(304,37, '%s' %label, color = 'red',size = 15)

for ii in range(len(exit_annot_list)):
    label = ('{}'.format(exit_annot_list[ii]))
    ax.text(exitPoint_horzAng[ii],exitPoint_vertAng[ii], '%s' %label, color = 'black',size = 10)
for ii in range(len(entry_annot_list)):
    label = ('{}'.format(entry_annot_list[ii]))
    ax.text(entryPoint_horzAng[ii],entryPoint_vertAng[ii], '%s' %label, color = 'red',size = 10)
plt.xlim([0,360])
plt.ylim([0,90])
plt.show()


check_fig5 = os.path.exists(('{0}/vert_VS_horz_{1}.png'.format(recording_path,recording_condition)))
if not check_fig5:
    fig5.savefig('{0}/vert_VS_horz_{1}.png'.format(recording_path,recording))

# Saving the exit and entry data

#### Saving the distance and angle of exit and entry points 

exitPoint_actDist_df = pd.DataFrame(exitPoint_actDist,columns=['distance_3D'])
exitPoint_horzAng_df = pd.DataFrame(exitPoint_horzAng,columns=['horz_angle'])
exitPoint_vertAng_df = pd.DataFrame(exitPoint_vertAng,columns=['vert_angle'])
exitDistAng_data = pd.concat([exitPoint_actDist_df,exitPoint_horzAng_df,exitPoint_vertAng_df],axis=1)
exitDistAng_data.head()

entryPoint_actDist_df = pd.DataFrame(entryPoint_actDist,columns=['distance_3D'])
entryPoint_horzAng_df = pd.DataFrame(entryPoint_horzAng,columns=['horz_angle'])
entryPoint_vertAng_df = pd.DataFrame(entryPoint_vertAng,columns=['vert_angle'])
entryDistAng_data = pd.concat([entryPoint_actDist_df,entryPoint_horzAng_df,entryPoint_vertAng_df],axis=1)
entryDistAng_data.head()

# saving the exit data
check_exit_data = os.path.exists('{0}/exitDistAng_data_{1}.csv'.format(recording_path,recording_condition))
if not check_exit_data:
    exitDistAng_data.to_csv('{0}/exitDistAng_data_{1}.csv'.format(recording_path,recording_condition))

# saving the entry data
check_entry_data = os.path.exists('{0}/entryDistAng_data_{1}.csv'.format(recording_path,recording_condition))
if not check_entry_data:
    entryDistAng_data.to_csv('{0}/entryDistAng_data_{1}.csv'.format(recording_path,recording_condition))
