==> 6_VRD_sorting_good_Vs_bad_recordings.ipynb <==
# Sort the learning and test flight data_exclude bad data

import os
import pandas as pd
import numpy as np
import math

# matplot
import matplotlib
import matplotlib.pyplot as plt

from matplotlib.projections import PolarAxes
from mpl_toolkits.axisartist.floating_axes import GridHelperCurveLinear, FloatingSubplot
import mpl_toolkits.axisartist.grid_finder as gf

# Config
import configparser
import plotly.graph_objects as go

# common_path
common_path = '/path/'


# selected good recording 
recording_df = pd.read_excel('{}/recordings.xlsx'.format(common_path))


recording_df.head()

# All bee list
bee_ids_list = list(recording_df.BeeID)

# removing duplicate bee ids
bee_ids_list = list(dict.fromkeys(bee_ids_list)) # remove the duplicate


print(bee_ids_list)
print(len(bee_ids_list))

# sorting the bee ids
bee_ids_list.sort()
print(bee_ids_list)
print(len(bee_ids_list))


# making the empty list with column names 
lf_Reco_sorted = pd.DataFrame(columns=bee_ids_list)
tf_Reco_sorted = pd.DataFrame(columns=bee_ids_list)

# adding sub-headers
tf_Reco_sorted


lf_Reco_sorted

recording_df.head()

#### Sort the good recordings 

# filling the empty data frame, learning flight and consecutive test flight based on bee id

for ii in range(len(bee_ids_list)): # for all the bee ids
    bee_id = bee_ids_list[ii] # selecting a bee
    bee_data = recording_df.loc[recording_df['BeeID']==bee_id] # getting all the data of the bee
    
    idx = 0 
    for a in range(0,len(bee_data),2): 
                
        if bee_data.quality.iloc[a] == 'good':  # quality manualy added, if it is good
            
            file_LF = bee_data.Filename.iloc[a] # seleting the learning recording 
            lf_Reco_sorted.at[idx,bee_id] = file_LF # adding the recording in the empty data frame of learning flight
            
            if bee_data.quality.iloc[a+1] == 'good':
                file_TF = bee_data.Filename.iloc[a+1] # selecting the test flight recording 
                tf_Reco_sorted.at[idx,bee_id] = file_TF # adding the recording in the empty data frame of test flight
            
            elif bee_data.quality.iloc[a+1] == 'not_good':
                tf_Reco_sorted.at[idx,(bee_id)] = np.nan # add NaN
            
            idx = idx+1
        

        elif bee_data.quality.iloc[a] == 'not_good':  # if data is not good
            lf_Reco_sorted.at[idx,(bee_id)] = np.nan # add NaN
            
            if bee_data.quality.iloc[a+1] == 'good':
                file_TF = bee_data.Filename.iloc[a+1] # selecting the test flight recording 
                tf_Reco_sorted.at[idx,bee_id] = file_TF # adding the recording in the empty data frame of test flight

            elif bee_data.quality.iloc[a+1] == 'not_good':
                tf_Reco_sorted.at[idx,(bee_id)] = np.nan # add NaN
            
            idx = idx+1


There are two conditions for bad quality, 
1. bad recoding 
2. only learning is recorded and not the test flight (inturpted learning flight)

# learning flight recordings - sorted
lf_Reco_sorted 

# Test flight recordings - sorted
tf_Reco_sorted

check_lfReco_sorted = os.path.exists('{}/lf_Reco_sorted.csv'.format(common_path))
if not check_lfReco_sorted:
    lf_Reco_sorted.to_csv('{}/lf_Reco_sorted.csv'.format(common_path),index=False)


check_tfReco_sorted = os.path.exists('{}/tf_Reco_sorted.csv'.format(common_path))
if not check_tfReco_sorted:
    tf_Reco_sorted.to_csv('{}/tf_Reco_sorted.csv'.format(common_path),index=False)
